refactor one monster test suite into three suites
