+++ tla-io/src/test/scala/at/forsyte/apalache/tla/imp/TestSanyImporter.scala
@@ -2,17 +2,13 @@ package at.forsyte.apalache.tla.imp
 
 import java.io.{File, PrintWriter}
 import java.nio.file.Files
-import at.forsyte.apalache.tla.imp.src.SourceStore
 import at.forsyte.apalache.tla.lir.convenience.tla._
 import at.forsyte.apalache.tla.lir.oper._
 import at.forsyte.apalache.tla.lir.src.{SourcePosition, SourceRegion}
 import at.forsyte.apalache.tla.lir.values._
 import at.forsyte.apalache.tla.lir.{NameEx, OperEx, ValEx, _}
-import at.forsyte.apalache.io.annotations.store._
 import at.forsyte.apalache.tla.lir.UntypedPredefs._
 import org.junit.runner.RunWith
-import org.scalactic.source.Position
-import org.scalatest.{BeforeAndAfter, FunSuite}
 import org.scalatest.junit.JUnitRunner
 
 import scala.collection.immutable.HashSet
@@ -24,85 +20,7 @@ import scala.io.Source
  * @author Igor Konnov
  */
 @RunWith(classOf[JUnitRunner])
-class TestSanyImporter extends FunSuite with BeforeAndAfter {
-  private var sourceStore: SourceStore = _
-  private var annotationStore: AnnotationStore = _
-  private var sanyImporter: SanyImporter = _
-
-  before {
-    sourceStore = new SourceStore
-    annotationStore = createAnnotationStore()
-    sanyImporter = new SanyImporter(sourceStore, annotationStore)
-  }
-
-  // assert that all declarations have source information
-  def expectSourceInfoInDefs(module: TlaModule): Unit = {
-    def assertForDecl(decl: TlaDecl): Unit = {
-      assert(
-          sourceStore.contains(decl.ID),
-          s"(No source location for declaration $decl.name)",
-      )
-    }
-
-    def collectDefs: TlaEx => Seq[TlaDecl] = {
-      case LetInEx(body, defs @ _*) =>
-        collectDefs(body) ++ defs.flatMap { d =>
-          collectDefs(d.body)
-        }
-
-      case OperEx(_, args @ _*) =>
-        args flatMap collectDefs
-
-      case _ =>
-        List()
-    }
-
-    for (decl <- module.declarations) {
-      assertForDecl(decl)
-      decl match {
-        case operDecl: TlaOperDecl =>
-          collectDefs(operDecl.body) foreach assertForDecl
-
-        case _ => ()
-      }
-    }
-  }
-
-  def expectOperDecl(
-      name: String, params: List[OperParam], body: TlaEx,
-  ): (TlaDecl => Unit) = {
-    case d: TlaOperDecl =>
-      assert(name == d.name)
-      assert(params == d.formalParams)
-      assert(body == d.body)
-      // the source location of the definition body has been saved
-      assert(
-          sourceStore.contains(d.body.ID),
-          s"(Source location of the definition body ${d.body.ID} is not stored)",
-      )
-      // the source location of the definition has been saved
-      assert(
-          sourceStore.contains(d.ID),
-          s"(Source of the definition ${d.name} is not stored)",
-      )
-
-    case d @ _ =>
-      fail("Expected a TlaOperDecl, found: " + d)
-  }
-
-  def findAndExpectOperDecl(
-      mod: TlaModule, name: String, params: List[OperParam], body: TlaEx,
-  ): Unit = {
-    mod.declarations.find {
-      _.name == name
-    } match {
-      case Some(d: TlaOperDecl) =>
-        expectOperDecl(name, params, body)(d)
-
-      case _ =>
-        fail("Expected a TlaDecl")
-    }
-  }
+class TestSanyImporter extends SanyImporterTestBase {
 
   test("empty module") {
     val text =
@@ -1085,27 +1003,6 @@ class TestSanyImporter extends FunSuite with BeforeAndAfter {
             ),
         ),
     )(mod.declarations(4))
-
-    /*  // the old test when Desugarer was part of SanyImporter
-    expectDecl("ExceptManyAt",
-      List(),
-      OperEx(TlaFunOper.except,
-        NameEx("x"),
-        TlaFunOper.mkTuple(ValEx(TlaInt(1))),
-        OperEx(TlaFunOper.except,
-          OperEx(TlaFunOper.app,
-            NameEx("x"),
-            ValEx(TlaInt(1))),
-          TlaFunOper.mkTuple(ValEx(TlaInt(2))),
-          OperEx(TlaBoolOper.and,
-            OperEx(TlaFunOper.app,
-              OperEx(TlaFunOper.app,
-                NameEx("x"), ValEx(TlaInt(1))),
-              ValEx(TlaInt(2))
-            ), ///
-            ValEx(TlaBool(true)))
-        )))(mod.declarations(4))
-     */
   }
 
   test("expression labels") {
@@ -1372,18 +1269,6 @@ class TestSanyImporter extends FunSuite with BeforeAndAfter {
             name("Z"),
         ),
     )(mod.declarations(5))
-
-    // the old test when Desugarer was part of SanyImporter
-    /*
-    expectDecl("C4",
-      OperEx(TlaFunOper.funDef,
-        ValEx(TlaBool(true)),
-        NameEx("x"), NameEx("X"),
-        NameEx("y"), NameEx("X"),
-        NameEx("a_b"), NameEx("Z"), // the tuple <<a, b>> is collapsed to a_b by Desugarer
-        NameEx("z"), NameEx("Z")
-      ))(mod.declarations(5))
-     */
   }
 
   test("weird set comprehension") {
@@ -1825,862 +1710,6 @@ class TestSanyImporter extends FunSuite with BeforeAndAfter {
     )
   }
 
-  test("instances") {
-    val text =
-      """---- MODULE inst ----
-        |---- MODULE A ----
-        |EXTENDS Naturals
-        |CONSTANT N
-        |F(x) == x + N
-        |G == F(3)
-        |================================
-        |CONSTANT M
-        |J == INSTANCE A WITH N <- M
-        |\* I(V, K) == INSTANCE A WITH N <- K
-        |---- MODULE B ----
-        |---- MODULE C ----
-        |H(x) == x
-        |================================
-        |CONSTANT M
-        |C2 == INSTANCE C
-        |F(x) == C2!H(x)
-        |================================
-        |B2 == INSTANCE B
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("inst", Source.fromString(text))
-    assert(2 == modules.size) // inst + Naturals
-    // the root module and A
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx): Unit =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    // the root module contains its own declarations and the declarations by FiniteSets
-    root.declarations.find { _.name == "J!F" } match {
-      case Some(
-              TlaOperDecl(
-                  _,
-                  params,
-                  OperEx(TlaArithOper.plus, NameEx("x"), NameEx("M")),
-              ),
-          ) =>
-        assert(params.length == 1)
-        assert(params.head.isInstanceOf[OperParam])
-        assert("x" == params.head.asInstanceOf[OperParam].name)
-
-      case e =>
-        fail("expected the body for J!F, found: " + e)
-    }
-    val fDecl = root.declarations.find { _.name == "J!F" }.get
-    root.declarations.find { _.name == "J!G" } match {
-      case Some(TlaOperDecl(_, params, body)) =>
-        assert(params.isEmpty)
-        val expected: TlaEx = appDecl(fDecl.asInstanceOf[TlaOperDecl], int(3))
-        assert(body == expected)
-
-      case _ =>
-        fail("expected the body for J!G")
-    }
-    // an instance with parameters, not working yet
-    /*
-    root.declarations.find { _.name == "I!F" } match {
-      case Some(TlaOperDecl(_, params,
-          OperEx(TlaArithOper.plus, NameEx("x"), NameEx("K")))) =>
-        assert(params.length == 3)
-        assert(params.head == FormalParam("V"))
-        assert(params(1) == FormalParam("K"))
-        assert(params(2) == FormalParam("x"))
-
-      case _ =>
-        fail("expected the body for I!F")
-    }
-     */
-    // two instances
-    root.declarations.find { _.name == "B2!C2!H" } match {
-      case Some(TlaOperDecl(_, params, NameEx("x"))) =>
-        assert(params.length == 1)
-        assert(params.head == OperParam("x"))
-
-      case _ =>
-        fail("expected the body for B2!C2!H")
-    }
-  }
-
-  test("EXTENDS of the standard modules") {
-    // operators with parameters that are themselves operators with parameters
-    val text =
-      """---- MODULE imports ----
-        |EXTENDS Naturals
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("imports", Source.fromString(text))
-    assert(2 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // we strip away the operator declarations by Naturals
-    assert(root.declarations.isEmpty)
-    // as Naturals containts definitions of the built-in operators, it is empty
-    val naturals = modules("Naturals")
-    assert(naturals.declarations.isEmpty)
-  }
-
-  test("lookup order in INSTANCE") {
-    // regression: SanyImporter was resolving the names in the wrong order
-    val text =
-      """---- MODULE imports ----
-        |---- MODULE M ----
-        |a == {}
-        |b == a
-        |================================
-        |a(self) == {self}
-        |I == INSTANCE M
-        |c == a(1)
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("imports", Source.fromString(text))
-    assert(1 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // we strip away the operator declarations by Naturals
-    assert(4 == root.declarations.size)
-    val aOfM = root.declarations.head
-    // check that the root module has indeed the definition a(self) == {self}, not a == {}
-    assert(
-        aOfM.isInstanceOf[TlaOperDecl] && aOfM
-          .asInstanceOf[TlaOperDecl]
-          .formalParams
-          .size == 1,
-    )
-  }
-
-  test("substitution in INSTANCE") {
-    // regression: the importer was not able to substitute a variable with an operator, e.g., in Paxos
-    val text =
-      """---- MODULE Paxos ----
-        |---- MODULE Consensus ----
-        |VARIABLE chosen
-        |C == chosen
-        |================================
-        |---- MODULE Voting ----
-        |chosen == 1
-        |C == INSTANCE Consensus
-        |================================
-        |V == INSTANCE Voting
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("Paxos", Source.fromString(text))
-    assert(1 == modules.size)
-    expectSourceInfoInDefs(modules(rootName))
-  }
-
-  test("LOCAL operator inside INSTANCE") {
-    val text =
-      """---- MODULE localInInstance ----
-        |---- MODULE M ----
-        |LOCAL a == {}
-        |b == a
-        |================================
-        |I == INSTANCE M
-        |c == I!b
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("localInInstance", Source.fromString(text))
-    assert(1 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // we strip away the operator declarations by Naturals
-    assert(2 == root.declarations.size)
-    val bOfM = root.declarations.head
-    // as "a" is LOCAL, it does not appear in the declarations, but it becomes a LET-IN definition inside b
-    assert("I!b" == bOfM.name)
-    // BUGFIX #576: use a unique lookup prefix to guarantee name uniqueness
-    bOfM match {
-      case TlaOperDecl(name, List(), LetInEx(OperEx(TlaOper.apply, NameEx(appliedName)), localDef)) =>
-        assert("I!b" == name)
-        assert(appliedName.startsWith("LOCAL"))
-        assert(appliedName.endsWith("!I!a"))
-        localDef match {
-          case TlaOperDecl(localName, List(), body) =>
-            assert(appliedName == localName)
-            assert(OperEx(TlaSetOper.enumSet) == body)
-
-          case _ =>
-            fail("Expected an operator definition " + appliedName)
-        }
-
-      case _ => fail("Unexpected operator structure")
-    }
-  }
-
-  // regression for #112
-  test("LET-IN inside INSTANCE") {
-    val text =
-      """---- MODULE letInInstance ----
-        |---- MODULE M ----
-        |VARIABLE x
-        |a ==
-        |  LET b == {} IN
-        |  b
-        |================================
-        |VARIABLE x
-        |INSTANCE M
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("letInInstance", Source.fromString(text))
-    assert(1 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // we strip away the operator declarations by Naturals
-    assert(2 == root.declarations.size)
-    val aOfM = root.declarations(1)
-    aOfM.asInstanceOf[TlaOperDecl].body match {
-      case LetInEx(body, bDecl) =>
-        assert(sourceStore.contains(body.ID))
-        assert(sourceStore.contains(bDecl.body.ID))
-
-      case _ =>
-        fail("expected declaration of b")
-    }
-  }
-
-  // regression for #143
-  test("Lookup inside substitutions") {
-    val text =
-      """------------------- MODULE P ----------------------
-        |------------------- MODULE Vot ----------------------
-        |------------------- MODULE Cons -------------------
-        |VARIABLE chosen
-        |Init == chosen = {}
-        |Next == chosen = {2}
-        |===================================================
-        |chosen == {2}
-        |C == INSTANCE Cons
-        |===================================================
-        |V == INSTANCE Vot
-        |===================================================""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("P", Source.fromString(text))
-    assert(1 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // expect V!chosen, V!C!Init and V!C!Next
-    assert(3 == root.declarations.size)
-    val next = root.declarations
-      .find(_.name == "V!C!Next")
-      .getOrElse(fail("V!C!Next not found"))
-    assert("V!C!Next" == next.name)
-    next.asInstanceOf[TlaOperDecl].body match {
-      case body @ OperEx(
-              TlaOper.eq,
-              OperEx(TlaOper.apply, NameEx("V!chosen")),
-              OperEx(TlaSetOper.enumSet, ValEx(TlaInt(i))),
-          ) =>
-        assert(i == 2)
-        assert(sourceStore.contains(body.ID))
-
-      case _ =>
-        fail("expected V!C!Next == V!chosen = {2}")
-    }
-  }
-
-  // regression for #143
-  test("Series of substitutions") {
-    val text =
-      """------------------- MODULE A ----------------------
-        |------------------- MODULE B ----------------------
-        |------------------- MODULE C -------------------
-        |VARIABLE x
-        |magic == x /= 2
-        |===================================================
-        |VARIABLE y
-        |C1 == INSTANCE C WITH x <- y
-        |===================================================
-        |VARIABLE z
-        |B1 == INSTANCE B WITH y <- z
-        |===================================================""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("A", Source.fromString(text))
-    assert(1 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // expect B1!C1!magic and z
-    assert(2 == root.declarations.size)
-    val magic = root.declarations
-      .find(_.name == "B1!C1!magic")
-      .getOrElse(fail("B1!C1!magic not found"))
-    assert("B1!C1!magic" == magic.name)
-    magic.asInstanceOf[TlaOperDecl].body match {
-      case body @ OperEx(TlaOper.ne, NameEx("z"), ValEx(TlaInt(i))) =>
-        assert(i == 2)
-        assert(sourceStore.contains(body.ID))
-
-      case _ =>
-        fail("expected B1!C1!magic == z /= 2")
-    }
-  }
-
-  // This test works due a temporary bugfix for issue #130.
-  // We should test it again, once the bug in SANY is fixed.
-  test("RECURSIVE operator inside INSTANCE") {
-    val text =
-      """---- MODULE recInInstance ----
-        |---- MODULE M ----
-        |RECURSIVE a(_)
-        |a(X) == a(X)
-        |================================
-        |RECURSIVE b(_)
-        |b(X) == b(X)
-        |I == INSTANCE M
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("recInInstance", Source.fromString(text))
-    assert(1 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // we strip away the operator declarations by Naturals
-    assert(2 == root.declarations.size)
-    val b = root.declarations.head
-    assert("b" == b.name)
-    assert(b.asInstanceOf[TlaOperDecl].isRecursive)
-    val aOfM = root.declarations(1)
-    // I!a should be recursive
-    assert("I!a" == aOfM.name)
-    assert(aOfM.asInstanceOf[TlaOperDecl].isRecursive)
-  }
-
-  test("module nats") {
-    // check that the Naturals module is imported properly
-    val text =
-      """---- MODULE nats ----
-        |EXTENDS Naturals
-        |NatSet == Nat
-        |Plus == 3 + 2
-        |Minus == 3 - 2
-        |Mult == 3 * 2
-        |Power == 3 ^ 2
-        |Less == 3 < 2
-        |Greater == 3 > 2
-        |Leq == 3 <= 2
-        |Leq2 == 3 =< 2
-        |Geq == 3 >= 2
-        |Mod == 3 % 2
-        |Div == 3 \div 2
-        |Range == 2 .. 3
-        |
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("nats", Source.fromString(text))
-    assert(2 == modules.size)
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx): Unit =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    // the root module contains its own declarations and the declarations by Naturals
-    expectDecl("NatSet", ValEx(TlaNatSet))
-    expectDecl(
-        "Plus",
-        OperEx(TlaArithOper.plus, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Minus",
-        OperEx(TlaArithOper.minus, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Mult",
-        OperEx(TlaArithOper.mult, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Power",
-        OperEx(TlaArithOper.exp, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Less",
-        OperEx(TlaArithOper.lt, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Greater",
-        OperEx(TlaArithOper.gt, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Leq",
-        OperEx(TlaArithOper.le, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Leq2",
-        OperEx(TlaArithOper.le, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Geq",
-        OperEx(TlaArithOper.ge, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Mod",
-        OperEx(TlaArithOper.mod, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Div",
-        OperEx(TlaArithOper.div, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Range",
-        OperEx(TlaArithOper.dotdot, ValEx(TlaInt(2)), ValEx(TlaInt(3))),
-    )
-
-    // check the source info
-    val plus = root.declarations.find {
-      _.name == "Plus"
-    } match {
-      case Some(TlaOperDecl(_, _, oe @ OperEx(oper, _*))) =>
-        val loc = sourceStore.find(oe.ID).get
-        assert(
-            SourceRegion(SourcePosition(4, 9), SourcePosition(4, 13)) == loc.region,
-        )
-
-      case _ => fail()
-    }
-  }
-
-  test("module ints") {
-    // check that the Integers module is imported properly
-    val text =
-      """---- MODULE ints ----
-        |EXTENDS Integers
-        |IntSet == Int
-        |Plus == 3 + 2
-        |Minus == 3 - 2
-        |Mult == 3 * 2
-        |Power == 3 ^ 2
-        |Less == 3 < 2
-        |Greater == 3 > 2
-        |Leq == 3 <= 2
-        |Geq == 3 >= 2
-        |Mod == 3 % 2
-        |Div == 3 \div 2
-        |Range == 2 .. 3
-        |UnaryMinus == -2
-        |
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("ints", Source.fromString(text))
-    assert(3 == modules.size) // Integers extends Naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx) =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    // the definitions of the standard operators are filtered out
-    assert(13 == root.declarations.size)
-
-    // the root module contains its own declarations and the declarations by Integers
-    expectDecl("IntSet", ValEx(TlaIntSet))
-    expectDecl(
-        "Plus",
-        OperEx(TlaArithOper.plus, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Minus",
-        OperEx(TlaArithOper.minus, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Mult",
-        OperEx(TlaArithOper.mult, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Power",
-        OperEx(TlaArithOper.exp, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Less",
-        OperEx(TlaArithOper.lt, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Greater",
-        OperEx(TlaArithOper.gt, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Leq",
-        OperEx(TlaArithOper.le, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Geq",
-        OperEx(TlaArithOper.ge, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Mod",
-        OperEx(TlaArithOper.mod, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Div",
-        OperEx(TlaArithOper.div, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Range",
-        OperEx(TlaArithOper.dotdot, ValEx(TlaInt(2)), ValEx(TlaInt(3))),
-    )
-    expectDecl("UnaryMinus", OperEx(TlaArithOper.uminus, ValEx(TlaInt(2))))
-  }
-
-  test("module reals") {
-    // check that the Reals module is imported properly
-    val text =
-      """---- MODULE reals ----
-        |EXTENDS Reals
-        |RealSet == Real
-        |Inf == Infinity
-        |Plus == 3 + 2
-        |Minus == 3 - 2
-        |Mult == 3 * 2
-        |Power == 3 ^ 2
-        |Less == 3 < 2
-        |Greater == 3 > 2
-        |Leq == 3 <= 2
-        |Geq == 3 >= 2
-        |Mod == 3 % 2
-        |Div == 3 \div 2
-        |Range == 2 .. 3
-        |UnaryMinus == -2
-        |RealDiv == 3 / 2
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("reals", Source.fromString(text))
-    assert(4 == modules.size) // Reals include Integers that include Naturals
-    val root = modules(rootName)
-    // the definitions of the standard operators are filtered out
-    assert(15 == root.declarations.size)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx) =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    expectDecl("RealSet", ValEx(TlaRealSet))
-    expectDecl("Inf", ValEx(TlaRealInfinity))
-    expectDecl(
-        "Plus",
-        OperEx(TlaArithOper.plus, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Minus",
-        OperEx(TlaArithOper.minus, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Mult",
-        OperEx(TlaArithOper.mult, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Power",
-        OperEx(TlaArithOper.exp, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Less",
-        OperEx(TlaArithOper.lt, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Greater",
-        OperEx(TlaArithOper.gt, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Leq",
-        OperEx(TlaArithOper.le, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Geq",
-        OperEx(TlaArithOper.ge, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Mod",
-        OperEx(TlaArithOper.mod, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Div",
-        OperEx(TlaArithOper.div, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-    expectDecl(
-        "Range",
-        OperEx(TlaArithOper.dotdot, ValEx(TlaInt(2)), ValEx(TlaInt(3))),
-    )
-    expectDecl("UnaryMinus", OperEx(TlaArithOper.uminus, ValEx(TlaInt(2))))
-    expectDecl(
-        "RealDiv",
-        OperEx(TlaArithOper.realDiv, ValEx(TlaInt(3)), ValEx(TlaInt(2))),
-    )
-  }
-
-  test("module seqs") {
-    // check that the Sequences module is imported properly
-    val text =
-      """---- MODULE sequences ----
-        |EXTENDS Sequences
-        |
-        |Empty == <<>>
-        |Three == <<1, 2, 3>>
-        |ASeq == Seq({1 , 2})
-        |ALen == Len(<<1, 2, 3>>)
-        |AConcat == <<1, 2>> \o <<3>>
-        |AAppend == Append(<<1, 2>>, 3)
-        |AHead == Head(<<1, 2, 3>>)
-        |ATail == Tail(<<1, 2, 3>>)
-        |ASubSeq == SubSeq(<<1, 2, 3, 4>>, 2, 3)
-        |Test(i) == i = 2
-        |ASelectSeq == SelectSeq(<<1, 2, 3, 4>>, Test)
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("sequences", Source.fromString(text))
-    assert(3 == modules.size) // Naturals, Sequences, and our module
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx) =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    // the definitions of the standard operators are filtered out
-    assert(11 == root.declarations.size)
-
-    expectDecl("Empty", tuple())
-    expectDecl("Three", tuple(int(1), int(2), int(3)))
-    expectDecl("ASeq", seqSet(enumSet(int(1), int(2))))
-    expectDecl("ALen", len(tuple(int(1), int(2), int(3))))
-    expectDecl("AConcat", concat(tuple(int(1), int(2)), tuple(int(3))))
-    expectDecl("AAppend", append(tuple(int(1), int(2)), int(3)))
-    expectDecl("AHead", head(tuple(int(1), int(2), int(3))))
-    expectDecl("ATail", tail(tuple(int(1), int(2), int(3))))
-    expectDecl(
-        "ASubSeq",
-        subseq(tuple(int(1), int(2), int(3), int(4)), int(2), int(3)),
-    )
-    expectDecl(
-        "ASelectSeq",
-        selectseq(tuple(int(1), int(2), int(3), int(4)), name("Test")),
-    )
-  }
-
-  // FiniteSets
-  test("module finitesets") {
-    // check that the FiniteSets module is imported properly
-    val text =
-      """---- MODULE finitesets ----
-        |EXTENDS FiniteSets
-        |IsFinSet == IsFiniteSet(BOOLEAN)
-        |Card == Cardinality(BOOLEAN)
-        |
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("finitesets", Source.fromString(text))
-    assert(4 == modules.size) // Naturals, Sequences, FiniteSets, and our module
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx) =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    // the definitions of the standard operators are filtered out
-    assert(2 == root.declarations.size)
-
-    // the root module contains its own declarations and the declarations by FiniteSets
-    expectDecl(
-        "IsFinSet",
-        OperEx(TlaFiniteSetOper.isFiniteSet, ValEx(TlaBoolSet)),
-    )
-    expectDecl("Card", OperEx(TlaFiniteSetOper.cardinality, ValEx(TlaBoolSet)))
-  }
-
-  test("module TLC") {
-    // check that the module TLC is imported properly
-    val text =
-      """---- MODULE tlc ----
-        |EXTENDS TLC
-        |
-        |VARIABLES i
-        |
-        |APrint == Print("TLC Print", TRUE)
-        |APrintT == PrintT("TLC PrintT")
-        |AAssert == Assert("TLC Assert", TRUE)
-        |AJavaTime == JavaTime
-        |ATLCGet == TLCGet(i)
-        |ATLCSet == TLCSet(i, 3)
-        |AColonGreater == {1, 2} :> 3
-        |AtAt == [j \in {1, 2} |-> j] @@ [k \in {3, 4} |-> k]
-        |APermutations == Permutations(<<1, 2>>)
-        |FakeSort(a, b) == TRUE
-        |ASortSeq == SortSeq(<<2, 1>>, FakeSort)
-        |ARandomElement == RandomElement({1, 2})
-        |AAny == Any
-        |AToString == ToString(42)
-        |ATLCEval == TLCEval(42)
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("tlc", Source.fromString(text))
-    // the root module and integers
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-
-    def expectDecl(name: String, body: TlaEx): Unit =
-      findAndExpectOperDecl(root, name, List(), body)
-
-    // the definitions of tlc and TLC
-    assert(30 == root.declarations.size)
-
-    expectDecl("APrint", OperEx(TlaOper.apply, NameEx("Print"), str("TLC Print"), bool(true)))
-    expectDecl("APrintT", OperEx(TlaOper.apply, NameEx("PrintT"), str("TLC PrintT")))
-    expectDecl("AAssert", OperEx(TlaOper.apply, NameEx("Assert"), str("TLC Assert"), bool(true)))
-    expectDecl("AJavaTime", OperEx(TlaOper.apply, NameEx("JavaTime")))
-    expectDecl("ATLCGet", OperEx(TlaOper.apply, NameEx("TLCGet"), name("i")))
-    expectDecl("ATLCSet", OperEx(TlaOper.apply, NameEx("TLCSet"), name("i"), int(3)))
-    expectDecl(
-        "AColonGreater",
-        OperEx(TlaOper.apply, NameEx(":>"), enumSet(int(1), int(2)), int(3)),
-    )
-    val fun12 = funDef(name("j"), name("j"), enumSet(int(1), int(2)))
-    val fun34 = funDef(name("k"), name("k"), enumSet(int(3), int(4)))
-    expectDecl("AtAt", OperEx(TlaOper.apply, NameEx("@@"), fun12, fun34))
-    expectDecl(
-        "APermutations",
-        OperEx(TlaOper.apply, NameEx("Permutations"), tuple(int(1), int(2))),
-    )
-    expectDecl(
-        "ASortSeq",
-        OperEx(TlaOper.apply, NameEx("SortSeq"), tuple(int(2), int(1)), name("FakeSort")),
-    )
-    expectDecl(
-        "ARandomElement",
-        OperEx(TlaOper.apply, NameEx("RandomElement"), enumSet(int(1), int(2))),
-    )
-    expectDecl("AAny", OperEx(TlaOper.apply, NameEx("Any")))
-    expectDecl("ATLCEval", OperEx(TlaOper.apply, NameEx("TLCEval"), int(42)))
-    expectDecl("AToString", OperEx(TlaOper.apply, NameEx("ToString"), int(42)))
-  }
-
-  test("EXTENDS Bags") {
-    // currently, Bags is imported as a user-defined module, no built-in operators are introduced
-    val text =
-      """---- MODULE localSum ----
-        |EXTENDS Bags
-        |================================
-      """.stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("localSum", Source.fromString(text))
-    assert(4 == modules.size) // Naturals, Sequences, TLC, FiniteSets, Bags, and our module
-
-    val root = modules("localSum")
-    expectSourceInfoInDefs(root)
-    // This number may change when a new version of Bags.tla is shipped in tla2tools.jar.
-    // The declarations include the declarations by __rewire_tlc_in_apalache.tla and Bags.tla.
-    assert(27 == root.declarations.size)
-  }
-
-  test("EXTENDS Apalache") {
-    val text =
-      """---- MODULE root ----
-        |EXTENDS Integers, FiniteSets, Apalache
-        |VARIABLE x, S
-        |
-        |Assn == x' := 1
-        |Sklm == Skolem(\E y \in S: TRUE)
-        |Expnd == Expand(SUBSET S)
-        |CC == ConstCardinality(Cardinality(S) >= 2)
-        |================================
-      """.stripMargin
-
-    // We have to set TLA-Library, in order to look up Apalache.tla. This is done automatically in pom.xml.
-    // If you run this test in an IDE, and the test fails, add the following line to the VM parameters
-    // (don't forget to replace <APALACHE_HOME> with the directory where you checked out the project):
-    //
-    // -DTLA-Library=<APALACHE_HOME>/src/tla
-    System.out.println(
-        "TLA-Library = %s".format(System.getProperty("TLA-Library")),
-    )
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("root", Source.fromString(text))
-    assert(6 == modules.size) // root, Apalache, Integers, FiniteSets, and whatever they import
-
-    val root = modules("root")
-    expectSourceInfoInDefs(root)
-    assert(6 == root.declarations.size)
-
-    def expectDecl(name: String, body: TlaEx) = {
-      findAndExpectOperDecl(root, name, List(), body)
-    }
-
-    expectDecl(
-        "Assn",
-        OperEx(
-            ApalacheOper.assign,
-            OperEx(TlaActionOper.prime, NameEx("x")),
-            ValEx(TlaInt(1)),
-        ),
-    )
-    expectDecl(
-        "Sklm",
-        OperEx(
-            ApalacheOper.skolem,
-            OperEx(
-                TlaBoolOper.exists,
-                NameEx("y"),
-                NameEx("S"),
-                ValEx(TlaBool(true)),
-            ),
-        ),
-    )
-    expectDecl(
-        "Expnd",
-        OperEx(ApalacheOper.expand, OperEx(TlaSetOper.powerset, NameEx("S"))),
-    )
-    expectDecl(
-        "CC",
-        OperEx(
-            ApalacheOper.constCard,
-            OperEx(
-                TlaArithOper.ge,
-                OperEx(TlaFiniteSetOper.cardinality, NameEx("S")),
-                ValEx(TlaInt(2)),
-            ),
-        ),
-    )
-  }
-
   test("assumptions") {
     // checking that the assumptions are imported properly
     val text =
@@ -2714,84 +1743,6 @@ class TestSanyImporter extends FunSuite with BeforeAndAfter {
     // regression test for issue #25
     val names = HashSet(modules(rootName).assumeDeclarations.map(_.name): _*)
     assert(2 == names.size) // all assumptions must have unique names
-  }
-
-  test("instances of standard modules") {
-    // Issue #52: the built-in operators are not eliminated when using INSTANCE
-    val text =
-      """---- MODULE issue52 ----
-        |I == INSTANCE Sequences
-        |A == I!Append(<<>>, {})
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("issue52", Source.fromString(text))
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // the definitions of the standard operators are filtered out
-    assert(1 == root.declarations.size)
-    root.declarations(0) match {
-      case TlaOperDecl("A", _, body) =>
-        assert(append(tuple(), enumSet()).untyped() == body)
-
-      case d => fail("unexpected declaration: " + d)
-    }
-  }
-
-  test("operators of local instances of the standard modules") {
-    // Issue #295: the built-in operators are not eliminated when using LOCAL INSTANCE
-    val text =
-      """---- MODULE issue295 ----
-        |---- MODULE A ----
-        |LOCAL INSTANCE Sequences
-        |A == Append(<<>>, {})
-        |==================
-        |INSTANCE A
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("issue295", Source.fromString(text))
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // the definitions of the standard operators are filtered out
-    assert(1 == root.declarations.size)
-    root.declarations(0) match {
-      case TlaOperDecl("A", _, body) =>
-        assert(append(tuple(), enumSet()).untyped() == body)
-
-      case d => fail("unexpected declaration: " + d)
-    }
-  }
-
-  test("values of local instances of the standard modules") {
-    // Issue #295: the built-in operators are not eliminated when using LOCAL INSTANCE
-    val text =
-      """---- MODULE issue295 ----
-        |---- MODULE A ----
-        |LOCAL INSTANCE Integers
-        |A == Int
-        |==================
-        |INSTANCE A
-        |================================
-        |""".stripMargin
-
-    val (rootName, modules) = sanyImporter
-      .loadFromSource("issue295", Source.fromString(text))
-    // the root module and naturals
-    val root = modules(rootName)
-    expectSourceInfoInDefs(root)
-    // the definitions of the standard operators are filtered out
-    assert(1 == root.declarations.size)
-    root.declarations(0) match {
-      case TlaOperDecl("A", _, body) =>
-        assert(ValEx(TlaIntSet) == body)
-
-      case d => fail("unexpected declaration: " + d)
-    }
   }
 
   test("ignore theorems") {
